requirements.txt
dvc
scikit-learn
xgboost
pandas
pyyaml
joblib
matplotlib
xlrd